+++ ingenialink/ethercat/network.py
@@ -3,22 +3,63 @@ import sys
 import platform
 import subprocess
 import inspect
-from typing import List
-
-from ingenialink.network import NET_PROT
-from ingenialink.exceptions import ILFirmwareLoadError
-from ingenialink import bin as bin_module
+import time
+from collections import defaultdict
+from typing import Optional, Any, Callable, List, Dict
+from threading import Thread
 
 import ingenialogger
+import pysoem  # type: ignore
+
+from ingenialink.network import Network, NET_PROT, NET_STATE, NET_DEV_EVT
+from ingenialink.exceptions import ILFirmwareLoadError, ILError
+from ingenialink import bin as bin_module
+from ingenialink.ethercat.servo import EthercatServo
+from ingenialink.constants import DEFAULT_ECAT_CONNECTION_TIMEOUT
 
 logger = ingenialogger.get_logger(__name__)
 
 
-class EthercatNetwork:
+class NetStatusListener(Thread):
+    """Network status listener thread to check if the drive is alive.
+
+    Args:
+        network: Network instance of the EtherCAT communication.
+
+    """
+
+    def __init__(self, network: "EthercatNetwork", refresh_time: float = 0.25):
+        super(NetStatusListener, self).__init__()
+        self.__network = network
+        self.__refresh_time = refresh_time
+        self.__stop = False
+        self._ecat_master = self.__network._ecat_master
+
+    def run(self) -> None:
+        while not self.__stop:
+            self._ecat_master.read_state()
+            for servo in self.__network.servos:
+                slave_id = servo.slave_id
+                servo_state = self.__network._get_servo_state(slave_id)
+                slave_is_alive = servo.slave.state == pysoem.NONE_STATE
+                if servo_state == NET_STATE.CONNECTED and slave_is_alive:
+                    self.__network._notify_status(slave_id, NET_DEV_EVT.REMOVED)
+                    self.__network._set_servo_state(slave_id, NET_STATE.DISCONNECTED)
+                if servo_state == NET_STATE.DISCONNECTED and not slave_is_alive:
+                    self.__network._notify_status(slave_id, NET_DEV_EVT.ADDED)
+                    self.__network._set_servo_state(slave_id, NET_STATE.CONNECTED)
+                time.sleep(self.__refresh_time)
+
+    def stop(self) -> None:
+        self.__stop = True
+
+
+class EthercatNetwork(Network):
     """Network for all EtherCAT communications.
 
     Args:
         interface_name: Interface name to be targeted.
+        connection_timeout: Time in seconds of the connection timeout.
 
     """
 
@@ -32,14 +73,126 @@ class EthercatNetwork:
     }
     UNKNOWN_FOE_ERROR = "Unknown error"
 
-    def __init__(self, interface_name: str) -> None:
-        self.interface_name = interface_name
-        """Interface name used in the network settings."""
-        self.servos: List[int] = []
-        """List of the connected servos in the network."""
-        self._ecat_master = None
+    def __init__(
+        self, interface_name: str, connection_timeout: float = DEFAULT_ECAT_CONNECTION_TIMEOUT
+    ):
+        super(EthercatNetwork, self).__init__()
+        self.interface_name: str = interface_name
+        self.servos: List[EthercatServo] = []
+        self.__servos_state: Dict[int, NET_STATE] = {}
+        self.__listener_net_status: Optional[NetStatusListener] = None
+        self.__observers_net_state: Dict[int, List[Any]] = defaultdict(list)
+        self._ecat_master: pysoem.CdefMaster = pysoem.Master()
+        self._ecat_master.sdo_read_timeout = int(1_000_000 * connection_timeout)
+        self._ecat_master.sdo_write_timeout = int(1_000_000 * connection_timeout)
+        self.__is_master_running = False
 
-    def load_firmware(self, fw_file: str, slave_id: int = 1) -> None:
+    def scan_slaves(self) -> List[int]:
+        """Scans for nodes in the network.
+
+        Returns:
+            List containing all the detected node IDs.
+
+        """
+        self._start_master()
+        nodes = self._ecat_master.config_init()
+        return list(range(1, nodes + 1))
+
+    def connect_to_slave(  # type: ignore [override]
+        self,
+        slave_id: int,
+        dictionary: str,
+        servo_status_listener: bool = False,
+        net_status_listener: bool = False,
+    ) -> EthercatServo:
+        """Connects to a drive through a given slave number.
+
+        Args:
+            slave_id: Targeted slave to be connected.
+            dictionary: Path to the dictionary file.
+            servo_status_listener: Toggle the listener of the servo for
+                its status, errors, faults, etc.
+            net_status_listener: Toggle the listener of the network
+                status, connection and disconnection.
+
+        Raises:
+            ValueError: If the slave ID is not valid.
+            ILError: If no slaves are found.
+
+        """
+        if not isinstance(slave_id, int) or slave_id < 0:
+            raise ValueError("Invalid slave ID value")
+        slaves = self.scan_slaves()
+        if len(slaves) == 0:
+            raise ILError("Could not find any slaves in the network.")
+        if slave_id not in slaves:
+            raise (ILError(f"Slave {slave_id} was not found."))
+        slave = self._ecat_master.slaves[slave_id - 1]
+        servo = EthercatServo(slave, slave_id, dictionary, servo_status_listener)
+        self.servos.append(servo)
+        self._set_servo_state(slave_id, NET_STATE.CONNECTED)
+        if net_status_listener:
+            self.start_status_listener()
+        return servo
+
+    def disconnect_from_slave(self, servo: EthercatServo) -> None:  # type: ignore [override]
+        """Disconnects the slave from the network.
+
+        Args:
+            servo: Instance of the servo connected.
+
+        """
+        self.servos.remove(servo)
+        if not self.servos:
+            self.stop_status_listener()
+            self._ecat_master.close()
+            self.__is_master_running = False
+
+    def subscribe_to_status(  # type: ignore [override]
+        self, slave_id: int, callback: Callable[[str, NET_DEV_EVT], None]
+    ) -> None:
+        """Subscribe to network state changes.
+
+        Args:
+            slave_id: Slave ID of the drive to subscribe.
+            callback: Callback function.
+
+        """
+        if callback in self.__observers_net_state[slave_id]:
+            logger.info("Callback already subscribed.")
+            return
+        self.__observers_net_state[slave_id].append(callback)
+
+    def unsubscribe_from_status(  # type: ignore [override]
+        self, slave_id: int, callback: Callable[[str, NET_DEV_EVT], None]
+    ) -> None:
+        """Unsubscribe from network state changes.
+
+        Args:
+            slave_id: Slave ID of the drive to subscribe.
+            callback: Callback function.
+
+        """
+        if callback not in self.__observers_net_state[slave_id]:
+            logger.info("Callback not subscribed.")
+            return
+        self.__observers_net_state[slave_id].remove(callback)
+
+    def start_status_listener(self) -> None:  # type: ignore [override]
+        """Start monitoring network events (CONNECTION/DISCONNECTION)."""
+        if self.__listener_net_status is None:
+            listener = NetStatusListener(self)
+            listener.start()
+            self.__listener_net_status = listener
+
+    def stop_status_listener(self) -> None:  # type: ignore [override]
+        """Stops the NetStatusListener from listening to the drive."""
+        if self.__listener_net_status is not None:
+            self.__listener_net_status.stop()
+            self.__listener_net_status.join()
+        self.__listener_net_status = None
+
+    def load_firmware(self, fw_file: str, slave_id: int = 1) -> None:  # type: ignore [override]
         """Loads a given firmware file to a target slave.
 
         Args:
@@ -80,7 +233,24 @@ class EthercatNetwork:
             ) from e
         logger.info("Firmware updated successfully")
 
+    def _start_master(self) -> None:
+        """Start the EtherCAT master if it has not been already started."""
+        if not self.__is_master_running:
+            self._ecat_master.open(self.interface_name)
+            self.__is_master_running = True
+
     @property
     def protocol(self) -> NET_PROT:
-        """Obtain network protocol."""
+        """NET_PROT: Obtain network protocol."""
         return NET_PROT.ECAT
+
+    def _get_servo_state(self, slave_id: int) -> NET_STATE:
+        return self.__servos_state[slave_id]
+
+    def _set_servo_state(self, slave_id: int, state: NET_STATE) -> None:
+        self.__servos_state[slave_id] = state
+
+    def _notify_status(self, slave_id: int, status: NET_DEV_EVT) -> None:
+        """Notify subscribers of a network state change."""
+        for callback in self.__observers_net_state[slave_id]:
+            callback(status)