+++ tests/ethercat/test_ethercat_network.py
@@ -1,7 +1,7 @@
 import pytest
 
 from ingenialink.ethercat.network import EthercatNetwork
-from ingenialink.exceptions import ILFirmwareLoadError
+from ingenialink.exceptions import ILFirmwareLoadError, ILError
 
 
 @pytest.mark.no_connection
@@ -23,16 +23,12 @@ def test_load_firmware_no_slave_detected_error(mocker, read_config):
 
 
 @pytest.mark.no_connection
-def test_load_firmware_value_error(mocker):
-    net = EthercatNetwork("not existing ifname")
-    mocker.patch("os.path.isfile", return_value=True)
-    with pytest.raises(
-        ILFirmwareLoadError,
-        match=(
-            "The firmware file could not be loaded correctly. Can’t initialize the network adapter"
-        ),
-    ):
-        net.load_firmware(fw_file="dummy_file.lfu")
+def test_wrong_interface_name_error(read_config):
+    with pytest.raises(ConnectionError):
+        net = EthercatNetwork("not existing ifname")
+        slave_id = 1
+        dictionary = read_config["ethernet"]["dictionary"]
+        net.connect_to_slave(slave_id, dictionary)
 
 
 @pytest.mark.no_connection
@@ -42,3 +38,23 @@ def test_load_firmware_not_implemented_error(mocker, read_config):
     mocker.patch("sys.platform", return_value="linux")
     with pytest.raises(NotImplementedError):
         net.load_firmware(fw_file="dummy_file.lfu")
+
+
+@pytest.mark.eoe
+@pytest.mark.parametrize("slave_id", [-1, "one", None])
+def test_connect_to_slave_invalid_id(read_config, slave_id):
+    net = EthercatNetwork(read_config["ethercat"]["ifname"])
+    with pytest.raises(ValueError):
+        net.connect_to_slave(slave_id)
+
+
+@pytest.mark.eoe
+def test_connect_to_slave_no_slaves_detected(mocker, read_config):
+    net = EthercatNetwork(read_config["ethercat"]["ifname"])
+
+    def scan_slaves():
+        return []
+
+    mocker.patch.object(net, "scan_slaves", scan_slaves)
+    with pytest.raises(ILError):
+        net.connect_to_slave(1)