+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Add
 - Support for multi-drive in CANopen's NetStatusListener.
+- EtherCAT communication via CoE (SDOs).
 - Add image attribute to dictionary class.
 
 ### Fixed
@@ -11,6 +12,9 @@
 - Set product name correctly if no dictionary is provided.
 - Docstrings from Register constructor and its subclasses are updated.
 - CanopenRegister and EthernetRegister have the same signature.
+
+### Changed
+- Raise ILValueError when the disturbance data does not fit in a register.
 
 ## [7.0.4] - 2023-10-11